Fix #1484: Markdown Editor
