+++ core/src/main/java/org/primefaces/extensions/component/markdowneditor/MarkdownEditor.java
@@ -23,9 +23,7 @@ package org.primefaces.extensions.component.markdowneditor;
 
 import javax.faces.application.ResourceDependency;
 
-import org.primefaces.component.api.AbstractPrimeHtmlInputTextArea;
-import org.primefaces.component.api.RTLAware;
-import org.primefaces.component.api.Widget;
+import org.primefaces.extensions.component.base.AbstractEditorInputTextArea;
 import org.primefaces.extensions.util.Constants;
 
 /**
@@ -39,7 +37,7 @@ import org.primefaces.extensions.util.Constants;
 @ResourceDependency(library = "primefaces", name = "core.js")
 @ResourceDependency(library = Constants.LIBRARY, name = "markdowneditor/markdowneditor.css")
 @ResourceDependency(library = Constants.LIBRARY, name = "markdowneditor/markdowneditor.js")
-public class MarkdownEditor extends AbstractPrimeHtmlInputTextArea implements Widget, RTLAware {
+public class MarkdownEditor extends AbstractEditorInputTextArea {
 
     public static final String COMPONENT_TYPE = "org.primefaces.extensions.component.MarkdownEditor";
     public static final String COMPONENT_FAMILY = "org.primefaces.extensions.component";
@@ -51,18 +49,7 @@ public class MarkdownEditor extends AbstractPrimeHtmlInputTextArea implements Wi
         minHeight,
         maxHeight,
         placeholder,
-        extender,
         sideBySideFullscreen,
-        allowBlocks,
-        allowFormatting,
-        allowLinks,
-        allowStyles,
-        allowImages,
-        allowTables,
-        allowMedia,
-        secure,
-        toolbar,
-        widgetVar
     }
     // @formatter:on
 
@@ -81,22 +68,6 @@ public class MarkdownEditor extends AbstractPrimeHtmlInputTextArea implements Wi
 
     public void setPlaceholder(String placeholder) {
         getStateHelper().put(PropertyKeys.placeholder, placeholder);
-    }
-
-    public String getWidgetVar() {
-        return (String) getStateHelper().eval(PropertyKeys.widgetVar, null);
-    }
-
-    public void setWidgetVar(String widgetVar) {
-        getStateHelper().put(PropertyKeys.widgetVar, widgetVar);
-    }
-
-    public String getExtender() {
-        return (String) getStateHelper().eval(PropertyKeys.extender, null);
-    }
-
-    public void setExtender(String extender) {
-        getStateHelper().put(PropertyKeys.extender, extender);
     }
 
     public String getMinHeight() {
@@ -123,75 +94,4 @@ public class MarkdownEditor extends AbstractPrimeHtmlInputTextArea implements Wi
         getStateHelper().put(PropertyKeys.sideBySideFullscreen, sideBySideFullscreen);
     }
 
-    public String getToolbar() {
-        return (String) getStateHelper().eval(PropertyKeys.toolbar, null);
-    }
-
-    public void setToolbar(String toolbar) {
-        getStateHelper().put(PropertyKeys.toolbar, toolbar);
-    }
-
-    public boolean isSecure() {
-        return (Boolean) getStateHelper().eval(PropertyKeys.secure, true);
-    }
-
-    public void setSecure(boolean secure) {
-        getStateHelper().put(PropertyKeys.secure, secure);
-    }
-
-    public boolean isAllowBlocks() {
-        return (Boolean) getStateHelper().eval(PropertyKeys.allowBlocks, true);
-    }
-
-    public void setAllowBlocks(boolean allowBlocks) {
-        getStateHelper().put(PropertyKeys.allowBlocks, allowBlocks);
-    }
-
-    public boolean isAllowFormatting() {
-        return (Boolean) getStateHelper().eval(PropertyKeys.allowFormatting, true);
-    }
-
-    public void setAllowFormatting(boolean allowFormatting) {
-        getStateHelper().put(PropertyKeys.allowFormatting, allowFormatting);
-    }
-
-    public boolean isAllowLinks() {
-        return (Boolean) getStateHelper().eval(PropertyKeys.allowLinks, true);
-    }
-
-    public void setAllowLinks(boolean allowLinks) {
-        getStateHelper().put(PropertyKeys.allowLinks, allowLinks);
-    }
-
-    public boolean isAllowStyles() {
-        return (Boolean) getStateHelper().eval(PropertyKeys.allowStyles, true);
-    }
-
-    public void setAllowStyles(boolean allowStyles) {
-        getStateHelper().put(PropertyKeys.allowStyles, allowStyles);
-    }
-
-    public boolean isAllowImages() {
-        return (Boolean) getStateHelper().eval(PropertyKeys.allowImages, true);
-    }
-
-    public void setAllowImages(boolean allowImages) {
-        getStateHelper().put(PropertyKeys.allowImages, allowImages);
-    }
-
-    public boolean isAllowTables() {
-        return (Boolean) getStateHelper().eval(PropertyKeys.allowTables, true);
-    }
-
-    public void setAllowMedia(boolean allowMedia) {
-        getStateHelper().put(PropertyKeys.allowMedia, allowMedia);
-    }
-
-    public boolean isAllowMedia() {
-        return (Boolean) getStateHelper().eval(PropertyKeys.allowMedia, true);
-    }
-
-    public void setAllowTables(boolean allowTables) {
-        getStateHelper().put(PropertyKeys.allowTables, allowTables);
-    }
 }
