+++ core/src/main/java/org/primefaces/extensions/component/suneditor/SunEditorRenderer.java
@@ -23,16 +23,12 @@ package org.primefaces.extensions.component.suneditor;
 
 import java.io.IOException;
 import java.util.Map;
-import java.util.logging.Logger;
 
-import javax.faces.FacesException;
 import javax.faces.component.UIComponent;
 import javax.faces.context.FacesContext;
 import javax.faces.context.ResponseWriter;
 import javax.faces.convert.Converter;
 
-import org.primefaces.context.PrimeApplicationContext;
-import org.primefaces.extensions.util.HtmlSanitizer;
 import org.primefaces.renderkit.InputRenderer;
 import org.primefaces.util.ComponentUtils;
 import org.primefaces.util.HTML;
@@ -46,8 +42,6 @@ import org.primefaces.util.WidgetBuilder;
  */
 public class SunEditorRenderer extends InputRenderer {
 
-    private static final Logger LOGGER = Logger.getLogger(SunEditorRenderer.class.getName());
-
     @Override
     public void decode(final FacesContext context, final UIComponent component) {
         final SunEditor editor = (SunEditor) component;
@@ -60,7 +54,7 @@ public class SunEditorRenderer extends InputRenderer {
         final String clientId = editor.getClientId(context);
         final Map<String, String> params = context.getExternalContext().getRequestParameterMap();
         if (params.containsKey(clientId)) {
-            String value = sanitizeHtml(context, editor, params.get(clientId));
+            String value = editor.sanitizeHtml(context, params.get(clientId));
             editor.setSubmittedValue(value);
         }
 
@@ -71,7 +65,7 @@ public class SunEditorRenderer extends InputRenderer {
     @Override
     public void encodeEnd(final FacesContext context, final UIComponent component) throws IOException {
         final SunEditor editor = (SunEditor) component;
-        checkSecurity(context, editor);
+        editor.checkSecurity(context);
         encodeMarkup(context, editor);
         encodeScript(context, editor);
     }
@@ -95,7 +89,7 @@ public class SunEditorRenderer extends InputRenderer {
         renderDomEvents(context, editor, HTML.INPUT_TEXT_EVENTS);
         renderValidationMetadata(context, editor);
 
-        String valueToRender = sanitizeHtml(context, editor, ComponentUtils.getValueToRender(context, editor));
+        String valueToRender = editor.sanitizeHtml(context, ComponentUtils.getValueToRender(context, editor));
         if (valueToRender != null) {
             writer.write(valueToRender);
         }
@@ -132,47 +126,5 @@ public class SunEditorRenderer extends InputRenderer {
         }
 
         return value;
-    }
-
-    /**
-     * Enforce security by default requiring the OWASP sanitizer on the classpath. Only if a user marks the editor with secure="false" will they opt-out of
-     * security.
-     *
-     * @param context the FacesContext
-     * @param editor the editor to check for security
-     */
-    private void checkSecurity(FacesContext context, SunEditor editor) {
-        boolean sanitizerAvailable = PrimeApplicationContext.getCurrentInstance(context).getEnvironment()
-                    .isHtmlSanitizerAvailable();
-        if (editor.isSecure() && !sanitizerAvailable) {
-            throw new FacesException(
-                        "SunEditor component is marked secure='true' but the HTML Sanitizer was not found on the classpath. "
-                                    + "Either add the HTML sanitizer to the classpath per the documentation"
-                                    + " or mark secure='false' if you would like to use the component without the sanitizer.");
-        }
-    }
-
-    /**
-     * If security is enabled sanitize the HTML string to prevent XSS.
-     *
-     * @param context the FacesContext
-     * @param editor the TextEditor instance
-     * @param value the value to sanitize
-     * @return the sanitized value
-     */
-    protected String sanitizeHtml(FacesContext context, SunEditor editor, String value) {
-        String result = value;
-        if (editor.isSecure()
-                    && PrimeApplicationContext.getCurrentInstance(context).getEnvironment().isHtmlSanitizerAvailable()) {
-            result = HtmlSanitizer.sanitizeHtml(value, editor.isAllowBlocks(), editor.isAllowFormatting(),
-                        editor.isAllowLinks(), editor.isAllowStyles(), editor.isAllowImages(), editor.isAllowTables(), editor.isAllowMedia());
-        }
-        else {
-            if (!editor.isAllowBlocks() || !editor.isAllowFormatting() || !editor.isAllowLinks()
-                        || !editor.isAllowStyles() || !editor.isAllowImages() || !editor.isAllowTables()) {
-                LOGGER.warning("HTML sanitizer not available - skip sanitizing....");
-            }
-        }
-        return result;
     }
 }
